refactor: remove tasks/ directory, consolidate into actions/builtins/
Move upload-cleanup, worktree-cleanup, log-cleanup from cron/tasks/
into cron/actions/builtins/ alongside their registration. Eliminates
the separate tasks/ layer — all action logic lives under actions/.
