+++ apps/api/src/cron/actions/builtins/worktree-cleanup.ts
