+++ apps/api/src/routes/settings/general.ts
@@ -17,7 +17,7 @@ import { DEFAULT_LOG_PAGE_SIZE, LOG_PAGE_SIZE_KEY } from '@/engines/issue/consta
 import { getCachedCategorizedCommands } from '@/engines/issue/queries'
 import type { WriteFilterRule } from '@/engines/write-filter'
 import { DEFAULT_FILTER_RULES, WRITE_FILTER_RULES_KEY } from '@/engines/write-filter'
-import { WORKTREE_AUTO_CLEANUP_KEY } from '@/cron/tasks/worktree-cleanup'
+import { WORKTREE_AUTO_CLEANUP_KEY } from '@/cron/actions/builtins/worktree-cleanup'
 
 const general = new Hono()
